[move] 플러그인 예시 router 파일 이동
user/user_router.py
