+++ plugin/demo_plugin/user/__init__.py
@@ -1,0 +1,7 @@
+from main import app
+from ..plugin_info import module_name, router_prefix
+from ..user.user_router import router
+
+
+def register_user_router():
+    app.include_router(router, prefix=f"/{router_prefix}", tags=[module_name])
