--- plugin/demo_plugin/user/router/__init__.py
@@ -1,7 +1,0 @@
-from main import app
-from ...plugin_info import module_name, router_prefix
-from .user_router import router
-
-
-def register_user_router():
-    app.include_router(router, prefix=f"/{router_prefix}", tags=[module_name])
