+++ plugin/demo_todo/templates/user_demo.html
@@ -1,0 +1,16 @@
+{% extends "base.html" %}
+
+{% block head %}
+{% endblock %}
+
+{% block title %}
+    {{ title }}
+{% endblock %}
+
+{% block subtitle %}
+    {{ title }}
+{% endblock %}
+
+{% block content %}
+    {{ content }}
+{% endblock %}
